ensembler/pyproject.toml
[tool.poetry]
name = "ensembler"
version = "0.1.0"
description = ""
authors = ["Anna Warno <awarno@7bulls.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"stomp.py" = "^8.0.0"
wheel = "^0.40.0"
sklearn = "^0.0"
pandas = "1.1.3"
pydantic = "1.8.2"
uvicorn = "0.15.0"
fastapi = "0.68.1"
cvxpy = "1.1.15"
filelock = "3.0.12"
influxdb = "^5.3.1"
python-slugify = "^6.1.1"
setproctitle = "^1.2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
